Added Context API for user data
